Prevent negative preroll values

+++ src/videosource.h
@@ -204,7 +204,7 @@ public:
     ~BestVideoSource();
     int GetTrack() const; // Useful when opening nth video track to get the actual number
     void SetMaxCacheSize(size_t Bytes); /* default max size is 1GB */
-    void SetSeekPreRoll(size_t Frames); /* the number of frames to cache before the position being fast forwarded to, default is 10 frames */
+    void SetSeekPreRoll(int64_t Frames); /* the number of frames to cache before the position being fast forwarded to, default is 10 frames */
     bool GetExactDuration();
     const VideoProperties &GetVideoProperties() const;
     BestVideoFrame *GetFrame(int64_t N);
